[241007] [refactor] 4.게시글 목록: 하위 컴포넌트 파일 분리
